+++ src/main/java/com/github/mavogel/ilias/state/states/action/GrantFileUploadToGroupMembersAction.java
@@ -28,6 +28,7 @@ import com.github.mavogel.client.ILIASSoapWebservicePortType;
 import com.github.mavogel.ilias.model.IliasNode;
 import com.github.mavogel.ilias.model.UserDataIds;
 import com.github.mavogel.ilias.state.ChangeAction;
+import com.github.mavogel.ilias.state.ToolStateMachine;
 import com.github.mavogel.ilias.utils.IOUtils;
 import com.github.mavogel.ilias.utils.IliasUtils;
 import org.apache.log4j.Logger;
@@ -35,6 +36,7 @@ import org.jdom.JDOMException;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Represents the action for granting the group member the permission to uplaod files.
@@ -46,7 +48,8 @@ public class GrantFileUploadToGroupMembersAction implements ChangeAction {
     private static Logger LOG = Logger.getLogger(GrantFileUploadToGroupMembersAction.class);
 
     @Override
-    public void performAction(final ILIASSoapWebservicePortType endpoint, final UserDataIds userDataIds, final List<IliasNode> nodes) {
+    public void performAction(final ILIASSoapWebservicePortType endpoint, Map<ToolStateMachine.ContextKey, List<IliasNode>> context,
+                              final UserDataIds userDataIds, final List<IliasNode> nodes) {
         LOG.info("Grant file upload to group members");
         if (confirm()) {
             final String sid = userDataIds.getSid();