+++ src/main/java/com/github/mavogel/ilias/state/states/action/SetRegistrationPeriodAction.java
@@ -30,6 +30,7 @@ import com.github.mavogel.ilias.model.IliasNode;
 import com.github.mavogel.ilias.model.RegistrationPeriod;
 import com.github.mavogel.ilias.model.UserDataIds;
 import com.github.mavogel.ilias.state.ChangeAction;
+import com.github.mavogel.ilias.state.ToolStateMachine;
 import com.github.mavogel.ilias.utils.IOUtils;
 import com.github.mavogel.ilias.utils.IliasUtils;
 import org.apache.log4j.Logger;
@@ -37,6 +38,7 @@ import org.jdom.JDOMException;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Map;
 
 /**
  * Represents the action for setting a new registration period.
@@ -48,8 +50,8 @@ public class SetRegistrationPeriodAction implements ChangeAction {
     private static Logger LOG = Logger.getLogger(SetRegistrationPeriodAction.class);
 
     @Override
-    public void performAction(final ILIASSoapWebservicePortType endpoint, final UserDataIds userDataIds,
-                                final List<IliasNode> nodes) {
+    public void performAction(final ILIASSoapWebservicePortType endpoint, Map<ToolStateMachine.ContextKey, List<IliasNode>> context,
+                              final UserDataIds userDataIds, final List<IliasNode> nodes) {
         LOG.info("Setting registration date");
         RegistrationPeriod registrationPeriod = IOUtils.readAndParseRegistrationDates();
         if (confirm()) {